fix(mcp): align entity schemas with handlers + serialize vectors (#130)
## Summary

Three latent bugs surfaced on the first real run of the v1 import →
analyze → set pipeline against a fresh user library:

1. **`entity_create(entity="track")`** — `TrackCreate` advertised
`provider_ids` while the `track_import` handler reads `external_ids`.
Following the schema produced `KeyError: 'external_ids'`. Default
`source` example pointed at `"yandex_music"` but `ProviderRegistry` only
has `"yandex"`.
2. **`entity_create(entity="audio_file")`** — schema offered `track_id`
(single) OR `track_ids` (batch); the handler only ever read `track_ids`,
so single-track callers hit `KeyError: 'track_ids'`.
3. **`entity_create(entity="track_features")`** — vector columns
(`mfcc_vector`, `tonnetz_vector`, `tempogram_ratio_vector`,
`beat_loudness_band_ratio`, `phrase_boundaries_ms`) are `Mapped[str |
None]` over `String(...)` in PG, but the analysis pipeline returns them
as `list[float]`. Every L3 analyze call crashed with `asyncpg DataError:
expected str, got list`.

This PR aligns the schemas with the actual handler contracts,
JSON-encodes vector lists transparently in `track_features.upsert`, and
adds 7 regression tests so the mismatch can't drift back in. The
expand-playlist workflow prompt was also pointing callers at the broken
API — updated to v1 surface.

Second commit (`chore`) is mypy-strict cleanup that `make check`
produces alongside the semantic changes — no behaviour delta.

## Test plan

- [x] `make check` — 729 passed (was 722, +7 regression tests), 44
xfailed, 20 xpassed
- [x] E2E pipeline on fresh track id 27271:
  - `entity_create(track, {source: yandex, external_ids: [85999444]})` ✅
  - `entity_create(audio_file, {track_id: 27271})` ✅ (single-form path)
- `entity_create(track_features, {track_ids: [27271], level: 3})` ✅ — 62
features, 0 errors
- [x] Full live run: 30 YM likes → import → download → L3 analyze →
set+set_version (set_id=46, quality 0.51 driven by source-data BPM
hopping, not transition math)


---------

+++ app/providers/yandex/rate_limiter.py
@@ -51,7 +51,7 @@ class TokenBucketRateLimiter:
     def current_backoff(self) -> float:
         if self._retry_count == 0:
             return 0.0
-        return self._base_backoff_s * (2 ** (self._retry_count - 1))
+        return float(self._base_backoff_s * (2 ** (self._retry_count - 1)))
 
     def retries_exhausted(self) -> bool:
         return self._retry_count >= self._max_retries
